Move definitions for template SLB_cin_obj.bt to a single file
